feat(rust): improve alignment of resolve behaviors with esbuild
