+++ crates/rolldown/src/utils/normalize_options.rs
@@ -1,5 +1,3 @@
-use rolldown_resolver::EnforceExtension;
-
 use crate::options::{
   normalized_input_options::NormalizedInputOptions,
   normalized_output_options::NormalizedOutputOptions, output_options::SourceMapType,
@@ -17,48 +15,9 @@ pub fn normalize_options(
   mut raw_input: crate::InputOptions,
   raw_output: crate::OutputOptions,
 ) -> NormalizeOptionsReturn {
-  let raw_resolve = std::mem::take(&mut raw_input.resolve).unwrap_or_default();
+  // Take out resolve options
 
-  // So far we align the default behavior with esbuild that target browser platform;
-  let resolve_options = rolldown_resolver::ResolveOptions {
-    tsconfig: None,
-    alias: raw_resolve
-      .alias
-      .map(|alias| {
-        alias
-          .into_iter()
-          .map(|(key, value)| {
-            (key, value.into_iter().map(rolldown_resolver::AliasValue::Path).collect::<Vec<_>>())
-          })
-          .collect::<Vec<_>>()
-      })
-      .unwrap_or_default(),
-    alias_fields: raw_resolve.alias_fields.unwrap_or_default(),
-    condition_names: raw_resolve.condition_names.unwrap_or_else(|| {
-      ["import", "default", "require"].into_iter().map(str::to_string).collect()
-    }),
-    description_files: vec!["package.json".to_string()],
-    enforce_extension: EnforceExtension::Auto,
-    exports_fields: raw_resolve.exports_fields.unwrap_or_else(|| vec![vec!["exports".to_string()]]),
-    extension_alias: vec![],
-    extensions: raw_resolve.extensions.unwrap_or_else(|| {
-      [".tsx", ".ts", ".jsx", ".js", ".json"].into_iter().map(str::to_string).collect()
-    }),
-    fallback: vec![],
-    fully_specified: false,
-    main_fields: raw_resolve
-      .main_fields
-      .unwrap_or_else(|| vec!["browser".to_string(), "module".to_string(), "main".to_string()]),
-    main_files: raw_resolve.main_files.unwrap_or_else(|| vec!["index".to_string()]),
-    modules: raw_resolve.modules.unwrap_or_else(|| vec!["node_modules".to_string()]),
-    resolve_to_context: false,
-    prefer_relative: false,
-    prefer_absolute: false,
-    restrictions: vec![],
-    roots: vec![],
-    symlinks: raw_resolve.symlinks.unwrap_or(true),
-    builtin_modules: false,
-  };
+  let raw_resolve = std::mem::take(&mut raw_input.resolve).unwrap_or_default();
 
   // Normalize input options
 
@@ -87,5 +46,5 @@ pub fn normalize_options(
     sourcemap: raw_output.sourcemap.unwrap_or(SourceMapType::Hidden),
   };
 
-  NormalizeOptionsReturn { input_options, output_options, resolve_options }
+  NormalizeOptionsReturn { input_options, output_options, resolve_options: raw_resolve }
 }
