+++ crates/rolldown_common/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-description = "This crate is mostly for sharing code between `rolldown` and `rolldown_plugin`"
+description = "This crate is mostly for sharing code between rolldwon crates."
 name        = "rolldown_common"
 version     = "0.0.1"
 
@@ -20,7 +20,6 @@ oxc                = { workspace = true, features = ["semantic"] }
 regex              = { workspace = true }
 rolldown_error     = { workspace = true }
 rolldown_fs        = { workspace = true }
-rolldown_resolver  = { workspace = true }
 rolldown_rstr      = { workspace = true }
 rolldown_sourcemap = { workspace = true }
 rustc-hash         = { workspace = true }
