+++ crates/rolldown_resolver/src/resolver.rs
@@ -1,22 +1,106 @@
+use rolldown_common::{ImportKind, ModuleType, Platform, ResolvedPath};
 use rolldown_error::BuildError;
 use rolldown_fs::FileSystem;
 use std::path::{Path, PathBuf};
 use sugar_path::SugarPathBuf;
 
-use oxc_resolver::{Resolution, ResolveError, ResolverGeneric};
+use oxc_resolver::{EnforceExtension, Resolution, ResolveError, ResolverGeneric};
 
-use crate::{types::resolved_path::ResolvedPath, ModuleType, ResolveOptions};
+use crate::types::resolve_options::ResolveOptions as NewResolveOptions;
+use oxc_resolver::ResolveOptions as OxcResolverOptions;
 
 #[derive(Debug)]
+#[allow(dead_code)]
 pub struct Resolver<T: FileSystem + Default> {
   cwd: PathBuf,
-  inner: ResolverGeneric<T>,
+  default_resolver: ResolverGeneric<T>,
+  import_resolver: ResolverGeneric<T>,
+  require_resolver: ResolverGeneric<T>,
 }
 
 impl<F: FileSystem + Default> Resolver<F> {
-  pub fn new(options: ResolveOptions, cwd: PathBuf, fs: F) -> Self {
-    let inner_resolver = ResolverGeneric::new_with_file_system(fs, options);
-    Self { cwd, inner: inner_resolver }
+  pub fn new(raw_resolve: NewResolveOptions, platform: Platform, cwd: PathBuf, fs: F) -> Self {
+    let mut default_conditions = vec!["default".to_string()];
+    let mut import_conditions = vec!["import".to_string()];
+    let mut require_conditions = vec!["require".to_string()];
+
+    default_conditions.extend(raw_resolve.condition_names.clone().unwrap_or_default());
+    match platform {
+      Platform::Node => {
+        default_conditions.push("node".to_string());
+      }
+      Platform::Browser => {
+        default_conditions.push("browser".to_string());
+      }
+      Platform::Neutral => {}
+    }
+    default_conditions.dedup();
+    import_conditions.extend(default_conditions.clone());
+    require_conditions.extend(default_conditions.clone());
+    import_conditions.dedup();
+    require_conditions.dedup();
+
+    let main_fields = raw_resolve.main_fields.clone().unwrap_or_else(|| match platform {
+      Platform::Node => {
+        vec!["main".to_string(), "module".to_string()]
+      }
+      Platform::Browser => vec!["browser".to_string(), "module".to_string(), "main".to_string()],
+      Platform::Neutral => vec![],
+    });
+
+    let resolve_options_with_default_conditions = OxcResolverOptions {
+      tsconfig: None,
+      alias: raw_resolve
+        .alias
+        .map(|alias| {
+          alias
+            .into_iter()
+            .map(|(key, value)| {
+              (key, value.into_iter().map(oxc_resolver::AliasValue::Path).collect::<Vec<_>>())
+            })
+            .collect::<Vec<_>>()
+        })
+        .unwrap_or_default(),
+      alias_fields: raw_resolve.alias_fields.unwrap_or_default(),
+      condition_names: default_conditions,
+      description_files: vec!["package.json".to_string()],
+      enforce_extension: EnforceExtension::Auto,
+      exports_fields: raw_resolve
+        .exports_fields
+        .unwrap_or_else(|| vec![vec!["exports".to_string()]]),
+      extension_alias: vec![],
+      extensions: raw_resolve
+        .extensions
+        .unwrap_or_else(|| [".jsx", ".js"].into_iter().map(str::to_string).collect()),
+      fallback: vec![],
+      fully_specified: false,
+      main_fields,
+      main_files: raw_resolve.main_files.unwrap_or_else(|| vec!["index".to_string()]),
+      modules: raw_resolve.modules.unwrap_or_else(|| vec!["node_modules".to_string()]),
+      resolve_to_context: false,
+      prefer_relative: false,
+      prefer_absolute: false,
+      restrictions: vec![],
+      roots: vec![],
+      symlinks: raw_resolve.symlinks.unwrap_or(true),
+      builtin_modules: false,
+    };
+    let resolve_options_with_import_conditions = OxcResolverOptions {
+      condition_names: import_conditions,
+      ..resolve_options_with_default_conditions.clone()
+    };
+    let resolve_options_with_require_conditions = OxcResolverOptions {
+      condition_names: require_conditions,
+      ..resolve_options_with_default_conditions.clone()
+    };
+    let default_resolver =
+      ResolverGeneric::new_with_file_system(fs, resolve_options_with_default_conditions);
+    let import_resolver =
+      default_resolver.clone_with_options(resolve_options_with_import_conditions);
+    let require_resolver =
+      default_resolver.clone_with_options(resolve_options_with_require_conditions);
+
+    Self { cwd, default_resolver, import_resolver, require_resolver }
   }
 
   pub fn cwd(&self) -> &PathBuf {
@@ -37,10 +121,15 @@ impl<F: FileSystem + Default> Resolver<F> {
     &self,
     importer: Option<&Path>,
     specifier: &str,
+    import_kind: ImportKind,
   ) -> Result<ResolveRet, BuildError> {
+    let selected_resolver = match import_kind {
+      ImportKind::Import | ImportKind::DynamicImport => &self.import_resolver,
+      ImportKind::Require => &self.require_resolver,
+    };
     let resolved = if let Some(importer) = importer {
       let context = importer.parent().expect("Should have a parent dir");
-      self.inner.resolve(context, specifier)
+      selected_resolver.resolve(context, specifier)
     } else {
       // If the importer is `None`, it means that the specifier is provided by the user in `input`. In this case, we can't call `resolver.resolve` with
       // `{ context: cwd, specifier: specifier }` due to rollup's default resolve behavior. For specifier `main`, rollup will try to resolve it as
@@ -51,12 +140,12 @@ impl<F: FileSystem + Default> Resolver<F> {
 
       let is_path_like = specifier.starts_with('.') || specifier.starts_with('/');
 
-      let resolved = self.inner.resolve(&self.cwd, joined_specifier.to_str().unwrap());
+      let resolved = self.default_resolver.resolve(&self.cwd, joined_specifier.to_str().unwrap());
       if resolved.is_ok() {
         resolved
       } else if !is_path_like {
         // If the specifier is not path-like, we should try to resolve it as a bare specifier. This allows us to resolve modules from node_modules.
-        self.inner.resolve(&self.cwd, specifier)
+        selected_resolver.resolve(&self.cwd, specifier)
       } else {
         resolved
       }
