+++ crates/rolldown_resolver/src/types/mod.rs
@@ -1,2 +1,1 @@
-pub mod module_type;
-pub mod resolved_path;
+pub mod resolve_options;
